move IdObjectIdGenerator and IdCacheResolver to tests as they don't take typename into account
