+++ apollo-normalized-cache-common/src/commonMain/kotlin/com/apollographql/apollo3/cache/normalized/ObjectIdGenerator.kt
@@ -39,17 +39,6 @@ class ObjectIdGeneratorContext(
 )
 
 /**
- * A [ObjectIdGenerator] that always uses the "id" field if it exists and delegates to [TypePolicyObjectIdGenerator] else
- *
- * It will coerce Int, Floats and other types to String using [toString]
- */
-object IdObjectIdGenerator : ObjectIdGenerator {
-  override fun cacheKeyForObject(obj: Map<String, Any?>, context: ObjectIdGeneratorContext): CacheKey? {
-    return obj["id"]?.toString()?.let { CacheKey(it) } ?: TypePolicyObjectIdGenerator.cacheKeyForObject(obj, context)
-  }
-}
-
-/**
  * A [ObjectIdGenerator] that uses annotations to compute the id
  */
 object TypePolicyObjectIdGenerator : ObjectIdGenerator {
