+++ composite/tests/integration-tests-kotlin/src/commonTest/kotlin/Utils.kt
@@ -1,3 +1,11 @@
+import com.apollographql.apollo3.api.CompiledField
+import com.apollographql.apollo3.api.Executable
+import com.apollographql.apollo3.cache.normalized.CacheKey
+import com.apollographql.apollo3.cache.normalized.CacheResolver
+import com.apollographql.apollo3.cache.normalized.FieldPolicyCacheResolver
+import com.apollographql.apollo3.cache.normalized.ObjectIdGenerator
+import com.apollographql.apollo3.cache.normalized.ObjectIdGeneratorContext
+import com.apollographql.apollo3.cache.normalized.TypePolicyObjectIdGenerator
 import com.apollographql.apollo3.testing.checkFile
 import com.apollographql.apollo3.testing.readFile
 import kotlin.test.assertEquals
@@ -12,3 +20,29 @@ fun readResource(name: String) = readFile("../integration-tests/testFixtures/res
  * A helper function to reverse the order of the argument so that we can easily column edit the tests
  */
 fun assertEquals2(actual: Any?, expected: Any?) = assertEquals(expected, actual)
+
+
+/**
+ * A [CacheResolver] that looks for an "id" argument to resolve fields and delegates to [FieldPolicyCacheResolver] else
+ */
+object IdCacheResolver: CacheResolver {
+  override fun resolveField(field: CompiledField, variables: Executable.Variables, parent: Map<String, Any?>, parentId: String): Any? {
+    val id = field.resolveArgument("id", variables)?.toString()
+    if (id != null) {
+      return CacheKey(id)
+    }
+
+    return FieldPolicyCacheResolver.resolveField(field, variables, parent, parentId)
+  }
+}
+
+/**
+ * A [ObjectIdGenerator] that always uses the "id" field if it exists and delegates to [TypePolicyObjectIdGenerator] else
+ *
+ * It will coerce Int, Floats and other types to String using [toString]
+ */
+object IdObjectIdGenerator : ObjectIdGenerator {
+  override fun cacheKeyForObject(obj: Map<String, Any?>, context: ObjectIdGeneratorContext): CacheKey? {
+    return obj["id"]?.toString()?.let { CacheKey(it) } ?: TypePolicyObjectIdGenerator.cacheKeyForObject(obj, context)
+  }
+}
